+++ CHANGELOG.rst
@@ -20,6 +20,8 @@ Changelog
 - Add 'obr apply' mode, see https://github.com/hpsim/OBR/pull/188
 - Add 'obr --version', see https://github.com/hpsim/OBR/pull/188
 - Add template block generators, see https://github.com/hpsim/OBR/pull/18://github.com/hpsim/OBR/pull/190 
+- Add 'validateState' operation, see https://github.com/hpsim/OBR/pull/189 
+- Improve 'run*Solver' launch speed, see https://github.com/hpsim/OBR/pull/189 
 
 
 0.2.0 (2023-09-14)